change to int64_t

+++ src/common/common.h
@@ -36,7 +36,7 @@ struct Value
     {
         int int_val;     // int value
         float float_val; // float value
-        long long bigint_val;
+        int64_t bigint_val;
     };
     bool invalid_val;
     std::string str_val; // string value
@@ -56,7 +56,7 @@ struct Value
         type = TYPE_INVALID;
         invalid_val = invalid_val_;
     }
-    void set_bigint(long long bigint_val_)
+    void set_bigint(int64_t bigint_val_)
     {
         type = TYPE_BIGINT;
         bigint_val = bigint_val_;
@@ -84,8 +84,8 @@ struct Value
         }
         else if (type == TYPE_BIGINT)
         {
-            assert(len == sizeof(long long));
-            *(long long *)(raw->data) = bigint_val;
+            assert(len == sizeof(int64_t));
+            *(int64_t *)(raw->data) = bigint_val;
         }
         else if (type == TYPE_INVALID)
         {
